requirements.txt
acres==0.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.9.5
aiohttp-retry==2.9.1
aiosignal==1.3.2
attrs==25.3.0
blinker==1.9.0
certifi>=2025.6.15
cffi>=1.17.1
charset-normalizer>=3.4.2
ci-info>=0.3.0
click>=8.2.1
configobj>=5.0.9
configparser>=7.2.0
cryptography>=45.0.4
etelemetry>=0.3.1
filelock>=3.18.0
fitz>=0.0.1
Flask>=3.1.1
frozenlist>=1.7.0
google-cloud-vision>=3.8.0
google-generativeai>=0.8.3
httplib2>=0.22.0
idna>=3.10
itsdangerous>=2.2.0
Jinja2>=3.1.6
looseversion>=1.3.0
lxml>=5.4.0
MarkupSafe>=3.0.2
multidict>=6.5.1
networkx>=3.5
nibabel>=5.3.2
nipype>=1.10.0
numpy>=2.3.1
opencv-python>=4.11.0.86
packaging>=25.0
pandas>=2.3.0
pathlib>=1.0.1
pdf2image>=1.17.0
pdfminer.six>=20250506
pdfplumber>=0.11.7
Pillow>=11.2.1
propcache>=0.3.2
prov>=2.1.1
puremagic>=1.29
pycparser>=2.22
pydot>=4.0.1
PyJWT>=2.10.1
PyMuPDF>=1.26.1
pyparsing>=3.2.3
pypdfium2>=4.30.1
python-dotenv>=1.0.0
pytesseract>=0.3.13
python-dateutil>=2.9.0
pytz>=2025.2
pyxnat>=1.6.3
rdflib>=7.1.4
requests>=2.32.4
scipy>=1.16.0
simplejson>=3.20.1
six>=1.17.0
traits>=7.0.2
twilio>=9.6.3
tzdata>=2025.2
urllib3>=2.5.0
Werkzeug>=3.1.3
yarl>=1.20.1
